seperate api and processing healthchecks
